+++ mosp/models/event.py
@@ -1,8 +1,7 @@
-from datetime import datetime
-
 from sqlalchemy.orm import validates
 
 from mosp.bootstrap import db
+from mosp.models._datetime import utcnow_naive
 
 
 class Event(db.Model):
@@ -13,7 +12,7 @@ class Event(db.Model):
     action = db.Column(db.String(), nullable=False)
     subject = db.Column(db.String(), nullable=False)
     initiator = db.Column(db.String())
-    date = db.Column(db.DateTime(), default=datetime.utcnow)
+    date = db.Column(db.DateTime(), default=utcnow_naive)
 
     @validates("initiator")
     def validates_initiator(self, key: str, value: str):