+++ mosp/models/license.py
@@ -1,6 +1,5 @@
-from datetime import datetime
-
 from mosp.bootstrap import db
+from mosp.models._datetime import utcnow_naive
 
 
 class License(db.Model):
@@ -11,7 +10,7 @@ class License(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(), default="", nullable=False, unique=True)
     license_id = db.Column(db.String(), default="", nullable=False, unique=True)
-    created_at = db.Column(db.DateTime(), default=datetime.utcnow)
+    created_at = db.Column(db.DateTime(), default=utcnow_naive)
 
     def __str__(self):
         return self.name