+++ mosp/views/session_mgmt.py
@@ -1,5 +1,4 @@
 import logging
-from datetime import datetime
 
 import sqlalchemy
 from flask import current_app
@@ -26,6 +25,7 @@ from mosp.bootstrap import db
 from mosp.forms import SigninForm
 from mosp.forms import SignupForm
 from mosp.models import User
+from mosp.models._datetime import utcnow_naive
 from mosp.notifications import notifications
 from mosp.views.common import admin_role
 from mosp.views.common import api_role
@@ -65,7 +65,7 @@ def load_user(user_id):
 @current_app.before_request
 def before_request():
     if current_user.is_authenticated:
-        current_user.last_seen = datetime.utcnow()
+        current_user.last_seen = utcnow_naive()
         db.session.commit()