+++ mosp/views/stats.py
@@ -1,6 +1,6 @@
-import datetime
 import re
 from collections import Counter
+from datetime import timedelta
 from urllib.parse import urljoin
 
 import networkx as nx
@@ -14,6 +14,7 @@ from mosp.bootstrap import application
 from mosp.models import Event
 from mosp.models import JsonObject
 from mosp.models import Schema
+from mosp.models._datetime import utcnow_naive
 
 
 stats_bp = Blueprint("stats_bp", __name__, url_prefix="/stats")
@@ -65,7 +66,7 @@ def digraph(software=None):
 @stats_bp.route("/objects/most-viewed.json", methods=["GET"])
 def most_viewed_objects():
     nb_weeks = request.args.get("nb_weeks", default=32, type=int)
-    nb_weeks_ago = datetime.datetime.utcnow() - datetime.timedelta(weeks=nb_weeks)
+    nb_weeks_ago = utcnow_naive() - timedelta(weeks=nb_weeks)
     events = Event.query.filter(
         Event.scope == "JsonObject",
         Event.action == "object_bp.view:GET",
@@ -112,7 +113,7 @@ def most_viewed_schemas():
 
     # look for uuid of schemas in JsonObject scope
     nb_weeks = request.args.get("nb_weeks", default=32, type=int)
-    nb_weeks_ago = datetime.datetime.utcnow() - datetime.timedelta(weeks=nb_weeks)
+    nb_weeks_ago = utcnow_naive() - timedelta(weeks=nb_weeks)
     events = Event.query.filter(
         Event.scope == "JsonObject",
         Event.action == "apiv2.object_objects_list:GET",