+++ src/codemirror-bundle.js
@@ -1,0 +1,2 @@
+export {EditorView, basicSetup} from "codemirror";
+export {json} from "@codemirror/lang-json";